pyproject.toml
[project]
name = "rawimg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=2.0.2",
    "pillow>=11.1.0",
    "pyobjc-core>=11.0",
    "pyobjc-framework-quartz>=11.0",
]
